feat: 중복코드를 최소화하도록 공통 컴포넌트 분리
