+++ src/components/Follower/Current/FollowerList.tsx
@@ -1,21 +1,11 @@
 import { useState } from 'react';
 import styled from 'styled-components';
 import CommonUserList from '../../../common/CommonUserList';
-import useGetFollowerSummary from '../../../libs/hooks/Follower/useGetFollowerSummary';
-import FollowerRecommendCard from '../Personal/FollowerRecommendCard';
 import FollowerFilter from './FollowerFilter';
 
 const FollowerList = () => {
   const [selectedGroupId, setSelectedGroupId] = useState(0);
   const [sorting, setSorting] = useState('최신순');
-  const [clickedPage, setClickedPage] = useState(1);
-
-  const { data, isLoading } = useGetFollowerSummary({
-    page: clickedPage - 1,
-    sortType: sorting,
-    groupId: selectedGroupId,
-  });
-  const { totalPage, followings } = !isLoading && data.data;
 
   const updateSelectedGroupId = (id: number) => {
     setSelectedGroupId(id);
@@ -28,18 +18,6 @@ const FollowerList = () => {
     setSorting(innerHTML);
   };
 
-  const handleClickPrevBtn = () => {
-    setClickedPage((prev) => prev - 1);
-  };
-
-  const handleClickPageNumber = (page: number) => {
-    setClickedPage(page);
-  };
-
-  const handleClickNextBtn = () => {
-    setClickedPage((prev) => prev + 1);
-  };
-
   return (
     <FollowerListContainer>
       <FollowerFilter
@@ -48,17 +26,7 @@ const FollowerList = () => {
         handleClickSorting={handleClickSorting}
       />
 
-      <CommonUserList
-        clickedPage={clickedPage}
-        totalPage={totalPage}
-        isLoading={isLoading}
-        followings={followings}
-        handleClickPrevBtn={handleClickPrevBtn}
-        handleClickPageNumber={handleClickPageNumber}
-        handleClickNextBtn={handleClickNextBtn}
-      />
-
-      {!isLoading && followings.length === 0 && <FollowerRecommendCard />}
+      <CommonUserList sorting={sorting} selectedGroupId={selectedGroupId} />
     </FollowerListContainer>
   );
 };
